refactor the current animal project logic to their own files
