+++ pkg/db/db_animalproject.go
@@ -1,9 +1,9 @@
 package db
 
 import (
-	"context"
-	"encoding/json"
-	"github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos"
+	_"context"
+	_ "encoding/json"
+	_ "github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos"
 )
 
 type Animal struct {
@@ -29,25 +29,6 @@ type AnimalProjectRelation struct {
 	ID string `json:"id"`
 	AnimalID string `json:"animalid"`
 	ProjectID string `json:"projectid"`
-	GenericDatabaseInfo
-}
-
-type Feed struct {
-	ID string `json:"id"`
-	Name string `json:"name"`
-	ProjectID string `json:"projectid"`
-	UserID string `json:"userid"`
-	GenericDatabaseInfo
-}
-
-type FeedPurchase struct {
-	ID string `json:"id"`
-	DatePurchased string `json:"date_purchased"`
-	AmountPurchased float64 `json:"amount_purchased"`
-	TotalCost float64 `json:"total_cost"`
-	FeedID string `json:"feedid"`
-	ProjectID string `json:"projectid"`
-	UserID string `json:"userid"`
 	GenericDatabaseInfo
 }
 
@@ -87,196 +68,6 @@ type Supplies struct {
 /*******************************
 * ANIMALS
 ********************************/
-
-/*******************************
-* FEED
-********************************/
-func (env *env) GetFeedsByProject(ctx context.Context, userid string, projectid string) ([]Feed, error) {
-
-	env.logger.Info("Getting feeds by project")
-
-	container, err := env.client.NewContainer("feeds")
-	if err != nil {
-		return []Feed{}, err
-	}
-
-	partitionKey := azcosmos.NewPartitionKeyString(userid)
-
-	query := "SELECT * FROM feeds f WHERE f.userid = @id AND f.projectid = @projectid"
-
-	queryOptions := azcosmos.QueryOptions{
-		QueryParameters: []azcosmos.QueryParameter{
-			{Name: "@id", Value: userid},
-			{Name: "@projectid", Value: projectid},
-		},
-	}
-
-	pager := container.NewQueryItemsPager(query, partitionKey, &queryOptions)
-
-	feeds := []Feed{}
-
-	for pager.More() {
-		response, err := pager.NextPage(ctx)
-		if err != nil {
-			return []Feed{}, err
-		}
-
-		for _, bytes := range response.Items {
-			feed := Feed{}
-			err := json.Unmarshal(bytes, &feed)
-			if err != nil {
-				return []Feed{}, err
-			}
-			feeds = append(feeds, feed)
-		}
-	}
-
-	return feeds, nil
-
-}
-
-func (env *env) GetFeedByID(ctx context.Context, userid string, feedid string) (Feed, error) {
-
-	env.logger.Info("Getting feed by ID")
-	feed := Feed{}
-
-	container, err := env.client.NewContainer("feeds")
-	if err != nil {
-		return feed, err
-	}
-
-	partitionKey := azcosmos.NewPartitionKeyString(userid)
-
-	response, err := container.ReadItem(ctx, partitionKey, feedid, nil)
-	if err != nil {
-		return feed, err
-	}
-
-	err = json.Unmarshal(response.Value, &feed)
-	if err != nil {
-		return feed, err
-	}
-
-	return feed, nil
-
-}
-
-func (env *env) UpsertFeed(ctx context.Context, feed Feed) (interface{}, error) {
-	
-	env.logger.Info("Upserting feed")
-
-	container, err := env.client.NewContainer("feeds")
-
-	partitionKey := azcosmos.NewPartitionKeyString(feed.UserID)
-
-	marshalled, err := json.Marshal(feed)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := container.UpsertItem(ctx, partitionKey, marshalled, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return response, nil
-
-}
-
-/*******************************
-* FEED PURCHASES
-********************************/
-func (env *env) GetFeedPurchasesByProject(ctx context.Context, userid string, projectid string) ([]FeedPurchase, error) {
-
-	env.logger.Info("Getting feed purchases by project")
-
-	container, err := env.client.NewContainer("feedpurchases")
-	if err != nil {
-		return []FeedPurchase{}, err
-	}
-
-	partitionKey := azcosmos.NewPartitionKeyString(userid)
-
-	query := "SELECT * FROM feedpurchases fp WHERE fp.userid = @id AND fp.projectid = @projectid"
-
-	queryOptions := azcosmos.QueryOptions{
-		QueryParameters: []azcosmos.QueryParameter{
-			{Name: "@id", Value: userid},
-			{Name: "@projectid", Value: projectid},
-		},
-	}
-
-	pager := container.NewQueryItemsPager(query, partitionKey, &queryOptions)
-
-	feedPurchases := []FeedPurchase{}
-
-	for pager.More() {
-		response, err := pager.NextPage(ctx)
-		if err != nil {
-			return []FeedPurchase{}, err
-		}
-
-		for _, bytes := range response.Items {
-			feedPurchase := FeedPurchase{}
-			err := json.Unmarshal(bytes, &feedPurchase)
-			if err != nil {
-				return []FeedPurchase{}, err
-			}
-			feedPurchases = append(feedPurchases, feedPurchase)
-		}
-	}
-
-	return feedPurchases, nil
-
-}
-
-func (env *env) GetFeedPurchaseByID(ctx context.Context, userid string, feedPurchaseID string) (FeedPurchase, error) {
-
-	env.logger.Info("Getting feed purchase by ID")
-	feedPurchase := FeedPurchase{}
-
-	container, err := env.client.NewContainer("feedpurchases")
-	if err != nil {
-		return feedPurchase, err
-	}
-
-	partitionKey := azcosmos.NewPartitionKeyString(userid)
-
-	response, err := container.ReadItem(ctx, partitionKey, feedPurchaseID, nil)
-	if err != nil {
-		return feedPurchase, err
-	}
-
-	err = json.Unmarshal(response.Value, &feedPurchase)
-	if err != nil {
-		return feedPurchase, err
-	}
-
-	return feedPurchase, nil
-
-}
-
-func (env *env) UpsertFeedPurchase(ctx context.Context, feedPurchase FeedPurchase) (interface{}, error) {
-	
-	env.logger.Info("Upserting feed purchase")
-
-	container, err := env.client.NewContainer("feedpurchases")
-
-	partitionKey := azcosmos.NewPartitionKeyString(feedPurchase.UserID)
-
-	marshalled, err := json.Marshal(feedPurchase)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := container.UpsertItem(ctx, partitionKey, marshalled, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return response, nil
-
-}
 
 /*******************************
 * DAILY FEED
